abstract executeCmd
