+++ src/main/utils/installExtensions.ts
@@ -1,0 +1,12 @@
+export const installExtensions = async () => {
+    const installer = require('electron-devtools-installer');
+    const forceDownload = !!process.env.UPGRADE_EXTENSIONS;
+    const extensions = ['REACT_DEVELOPER_TOOLS'];
+
+    return installer
+        .default(
+            extensions.map((name) => installer[name]),
+            forceDownload
+        )
+        .catch(console.log);
+};
