+++ src/main/utils/resolveHtmlPath.ts
@@ -15,16 +15,3 @@ if (process.env.NODE_ENV === 'development') {
         return `file://${path.resolve(__dirname, '../renderer/', htmlFileName)}`;
     };
 }
-
-export const installExtensions = async () => {
-    const installer = require('electron-devtools-installer');
-    const forceDownload = !!process.env.UPGRADE_EXTENSIONS;
-    const extensions = ['REACT_DEVELOPER_TOOLS'];
-
-    return installer
-        .default(
-            extensions.map((name) => installer[name]),
-            forceDownload
-        )
-        .catch(console.log);
-};
